+++ CotEditor/Sources/LineNumberView.swift
@@ -149,11 +149,21 @@ final class LineNumberView: NSRulerView {
     /// make background transparent
     override var isOpaque: Bool {
         
-        return self.textView?.isOpaque ?? true
+        guard let textView = self.textView else { return true }
+        
+        if textView.isOpaque { return true }
+        
+        // avoid overlapping line numbers with the text in the text view
+        // -> On macOS 10.14 (and later?), text view is drawn under the ruler view. (2018-11 macOS 10.14)
+        if NSAppKitVersion.current > .macOS10_13_4 {
+            return !textView.wrapsLines
+        }
+        
+        return false
     }
     
     
-    /// just before view will be attached
+    /// just before the receiver attaches to a superview
     override func viewWillMove(toSuperview newSuperview: NSView?) {
         
         super.viewWillMove(toSuperview: newSuperview)