+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Change Log
 ### Improvements
 
 - Avoid drawing variant sequence as invisible control characters.
+- Make line number view opaque if lines are unwrapped on macOS 10.14 to avoid drawing the editor's text over the line numbers.
 - [tirivial] Draw vertical tabs as general invisible control characters.